+++ apps/web/app/page.tsx
@@ -117,7 +117,7 @@ export default function Home() {
                   setLoading(false);
                 }}
               />
-              {loading && <Spinner color="secondary" />}
+              {loading && <Spinner color="default" />}
             </motion.div>
           </AnimatePresence>
           <div className="grid grid-cols-2 lg:grid-cols-4 w-full gap-6">