+++ apps/web/CHANGELOG.md
@@ -1,5 +1,33 @@
 # web
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [229fb29]
+  - @triyanox/react-video@0.1.5
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [05f3361]
+  - @triyanox/react-video@0.1.4
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [c4a2d4f]
+  - @triyanox/react-video@0.1.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [c0f5ed9]
+  - @triyanox/react-video@0.1.2
+
 ## 0.0.2
 
 ### Patch Changes