+++ packages/core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @triyanox/react-video
 
+## 0.1.5
+
+### Patch Changes
+
+- 229fb29: Remove double click to full screen
+
+## 0.1.4
+
+### Patch Changes
+
+- 05f3361: Hide fullscreen when pip
+
+## 0.1.3
+
+### Patch Changes
+
+- c4a2d4f: Fix framer-motion declarations problem
+
+## 0.1.2
+
+### Patch Changes
+
+- c0f5ed9: Fix framer-motion declarations problem
+
 ## 0.1.1
 
 ### Patch Changes